Moved image carosel modal into reusable component
